+++ bnb_(feat.flex_grid).css
@@ -295,7 +295,7 @@ header .basic_menu div {
 
 .toggleMenu {
   position: absolute;
-  right: 0;
+  right: 90px;
   background-color: rgba(255, 255, 255, 1);
   border-radius: 10px;
   padding: 10px 0px;
