--- checkInOut.js
@@ -1,13 +1,0 @@
-export class CheckInOut {
-   constructor(date = new Date()) {
-      this.date = date;
-
-   }
-   init() {
-      document.querySelectorAll(".calendar_current .calendar_date div").forEach(el => {
-         if (el.innerText < this.date.getDate()) {
-            el.classList.add("before_today");
-         }
-      })
-   }
-}
